Move standaloneTaskSid definition
