+++ plugin-hrm-form/src/states/ContactState.js
@@ -21,7 +21,8 @@ import callTypes from './DomainConstants';
 import { countSelectedCategories } from './ValidationRules';
 import { copySearchResultIntoTask } from './contacts/helpers';
 import { getConfig } from '../HrmFormPlugin';
-import { standaloneTaskSid } from './search/reducer';
+
+export const standaloneTaskSid = 'standalone-task-sid';
 
 /**
  * Looks for a particular task in the state object, and returns it if found.
