+++ plugin-hrm-form/src/states/search/reducer.ts
@@ -4,6 +4,7 @@ import * as t from './types';
 import { INITIALIZE_CONTACT_STATE, RECREATE_CONTACT_STATE, REMOVE_CONTACT_STATE, GeneralActionType } from '../types';
 import { SearchContact, SearchCaseResult } from '../../types/types';
 import { ContactDetailsSections, ContactDetailsSectionsType } from '../../components/common/ContactDetails';
+import { standaloneTaskSid } from '../ContactState';
 
 type TaskEntry = {
   currentPage: t.SearchPagesType;
@@ -51,8 +52,6 @@ export const newTaskEntry: TaskEntry = {
   error: null,
   casesError: null,
 };
-
-export const standaloneTaskSid = 'standalone-task-sid';
 
 export const initialState: SearchState = {
   tasks: {
